+++ CHANGES.rst
@@ -4,6 +4,8 @@ Next Release
 
 - Extend monthly membership invoice statistics by current balance.
 
+- Fix backend membership certificate url.
+
 
 
 1.12-beta.3